Changed the underlying storage for XmlField child objects
Changed the storage for XmlField::children data member from list to
multimap<string,XmlField>

Changed the XmlField::[] access operator to return a value without checking
for uniqueness or existence (for speed). Moved the checking version
to XmlField::get_field

Added a XmlField::get_field_range method to return an equal_range
of iterators pointing to all fields with given name.

Added a XmlField::fifo data member to store the iterators to child
fields in the order of input. Changed the output method to
traverse fifo instead of XmlField::children