Added path-validating features + minor convenience improvements
Added a XmlField::validate_path method (taking two list<string>
iterators) that can validate a pseudo-XPath relative to the top node,
and return an element at the end of this path.
Also added a field_exists check to validate immediate descendants of the
node.

Provide an const-const overload for the XmlField::[] single field access
operator. Provided a method that returns a (non-const) reference to
the attributes map of the XmlField object